refactor: tests of endpoints that could respond with meta data
